// Unicus Frontend/src/app/models/dto/fieldsToKeepDTO.ts
export class FieldsToKeepDTO {

  constructor() {}

  bsn: boolean = false;
  relatienummer: boolean = false;
  geslachtsnaam: boolean = false;
  voornamen: boolean = false;
  voorletters: boolean = false;
  tussenvoegsels: boolean = false;
  geslacht: boolean = false;
  geboortedatum: boolean = false;
  geboorteplaats: boolean  = false;
  geboorteland: boolean = false;
  nationaliteit: boolean = false;
  opmerking: boolean = false;
  adressen: boolean = false;
  partner: boolean = false;
  contactgegevens: boolean = false;
  betaalgegevens: boolean = false;
  labels: boolean = false;
  contactmomenten: boolean = false;
  identiteitsdocumenten: boolean = false;
}